Refactor: Extract PageToc component and use in blog and docs
